https://github.com/sdasgup3/validating-binary-decompilation/issues/16: SSB: WIP IV: MEM mismatch
For decompiling the called function, compd lifts, by default, a LLVM
declaration with function attributes like readnone returned to inform the
normalizer (which is the LLVM -03 opt) that an argument of the called function
is not going to changed and returned as is. This help the normalizer to
optimize better. The expectation is that the McSema lifted LLVM IR should hold
the same assumption of that called function.  For some function, the assuption
that the argument will remained unchanged is not true and to handle that we
need to pass special swich (--assume-none-decl-retval) to the compd tool to
avoid adding those function attributes. Helped in 104 cases.

            Later, found that different called function might have differnet
            assumptions about the returned value and hence we need each fnction
            call declaration to be trated separately in compd output. Hence
            decided to modify mcsema output by converting the called function
            defition to declaration.  This helps in handling ~403 more caes